rename file with kebab format
