+++ src/localization/translations/en/translation.ts
@@ -1,5 +1,5 @@
 import { common } from "./common";
-import { libraryPage } from "./libraryPage";
+import { libraryPage } from "./library-page";
 import { multimathPage } from "./multimath-page";
 import { welcomePage } from "./welcome-page";
